% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.R
\name{as.cluster.gce_instance}
\alias{as.cluster.gce_instance}
\title{future cluster for GCE objects}
\usage{
\method{as.cluster}{gce_instance}(x, user = gce_get_global_ssh_user(),
  project = gce_get_global_project(), zone = gce_get_global_zone(),
  rshopts = ssh_options(), ..., recursive = FALSE)
}
\arguments{
\item{x}{The instance to make a future cluster}

\item{user}{Username used in ssh}

\item{project}{The GCE project}

\item{zone}{The GCE zone}

\item{rshopts}{Options for the SSH}

\item{...}{other arguments passed to makeDockerClusterPSOCK}

\item{recursive}{recursive}
}
\description{
S3 method for as.cluster in the future package
}
\keyword{##}
\keyword{Imports:}
\keyword{REMOVE}
\keyword{future}
\keyword{iff}
\keyword{internal}

