% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.R
\name{as.cluster}
\alias{as.cluster}
\title{Make a cluster}
\usage{
as.cluster(x, ...)
}
\arguments{
\item{x}{Object to make a cluster passed to S3 method}

\item{...}{other arguments}
}
\description{
Used in future package
}
\keyword{internal}

