% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{as.container}
\alias{as.container}
\title{Coerce an object into a container object.}
\usage{
as.container(x, host = localhost)
}
\description{
Coerce an object into a container object.
}
\author{
Winston Chang \email{winston@stdout.org}
A container object represents a Docker container on a host.
}

