% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_admin.R
\name{as.gce_instance}
\alias{as.gce_instance}
\title{Turn an instance name into an instance, or return the instance}
\usage{
as.gce_instance(x, project = gce_get_global_project(),
  zone = gce_get_global_zone())
}
\arguments{
\item{x}{A character name of instance or instance object}

\item{project}{GDE project}

\item{zone}{GCE zone}
}
\description{
Turn an instance name into an instance, or return the instance
}
\keyword{internal}
