% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{as.gce_instance_name}
\alias{as.gce_instance_name}
\title{Get the instance name if passed an instance}
\usage{
as.gce_instance_name(x)
}
\arguments{
\item{a}{character name or gce_instance object}
}
\description{
Get the instance name if passed an instance
}
\keyword{internal}

