% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_admin.R
\name{as.gce_instance_name_one}
\alias{as.gce_instance_name_one}
\title{Get the instance name if passed an instance}
\usage{
as.gce_instance_name_one(x)
}
\arguments{
\item{a}{character name or gce_instance object}
}
\description{
Get the instance name if passed an instance
}
\keyword{internal}
