% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disks.R
\name{AttachedDiskInitializeParams}
\alias{AttachedDiskInitializeParams}
\title{AttachedDiskInitializeParams Object}
\usage{
AttachedDiskInitializeParams(diskName = NULL, diskSizeGb = NULL,
  diskType = NULL, sourceImage = NULL, sourceImageEncryptionKey = NULL)
}
\arguments{
\item{diskName}{Specifies the disk name}

\item{diskSizeGb}{Specifies the size of the disk in base-2 GB}

\item{diskType}{Specifies the disk type to use to create the instance}

\item{sourceImage}{The source image used to create this disk}

\item{sourceImageEncryptionKey}{The customer-supplied encryption key of the source image}
}
\value{
AttachedDiskInitializeParams object
}
\description{
AttachedDiskInitializeParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Input Only] Specifies the parameters for a new disk that will be created alongside the new instance. Use initialization parameters to create boot disks or local SSDs attached to the new instance.This property is mutually exclusive with the source property; you can only define one or the other, but not both.
}

