% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_admin.R
\name{cli_tools}
\alias{cli_tools}
\title{See if ssh or scp is installed
From https://github.com/sckott/analogsea/blob/master/R/zzz.R}
\usage{
cli_tools()
}
\description{
See if ssh or scp is installed
From https://github.com/sckott/analogsea/blob/master/R/zzz.R
}
\keyword{internal}
