% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_container.R
\name{container_rm}
\alias{container_rm}
\title{Delete a container.}
\usage{
container_rm(container, force = FALSE)
}
\arguments{
\item{container}{A container object}

\item{force}{Force removal of a running container.}
}
\description{
Delete a container.
}
\examples{
\dontrun{
container_rm(con)
}
}
\author{
Winston Change \email{winston@stdout.org}
}

