% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{container_update_info}
\alias{container_update_info}
\title{Update the information about a container.}
\usage{
container_update_info(container)
}
\description{
This queries docker (on the host) for information about the container, and
saves the returned information into a container object, which is returned.
This does not use reference semantics, so if you want to store the updated
information, you need to save the result.
}
\examples{
\dontrun{
con <- container_update_info(con)
}
}
\author{
Winston Chang \email{winston@stdout.org}
}

