% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{containers}
\alias{containers}
\title{Get list of all containers on a host.}
\usage{
containers(host = localhost, ...)
}
\description{
Get list of all containers on a host.
}
\author{
Winston Chang \email{winston@stdout.org}
}

