% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disks.R
\name{Disk}
\alias{Disk}
\title{Disk Object}
\usage{
Disk(name = NULL, sourceImage = NULL, sizeGb = NULL,
  description = NULL, diskEncryptionKey = NULL, licenses = NULL,
  sourceImageEncryptionKey = NULL, sourceSnapshot = NULL,
  sourceSnapshotEncryptionKey = NULL, type = NULL)
}
\arguments{
\item{name}{Name of the resource}

\item{sourceImage}{The source image used to create this disk}

\item{sizeGb}{Size of the persistent disk, specified in GB}

\item{description}{An optional description of this resource}

\item{diskEncryptionKey}{Encrypts the disk using a customer-supplied encryption key}

\item{licenses}{Any applicable publicly visible licenses}

\item{sourceImageEncryptionKey}{The customer-supplied encryption key of the source image}

\item{sourceSnapshot}{The source snapshot used to create this disk}

\item{sourceSnapshotEncryptionKey}{The customer-supplied encryption key of the source snapshot}

\item{type}{URL of the disk type resource describing which disk type to use to create the disk}
}
\value{
Disk object
}
\description{
Disk Object
}
\details{
A Disk resource.
}
\concept{Disk functions}
\keyword{internal}
