% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_gce.R
\name{docker_build}
\alias{docker_build}
\title{Build image on an instance from a local Dockerfile}
\usage{
docker_build(host = localhost, dockerfile, folder, new_image, ...)
}
\arguments{
\item{host}{A host object.}

\item{dockerfile}{Location of local dockerfile}

\item{folder}{Where on host to build dockerfile}

\item{new_image}{Name of the new image}

\item{...}{Other arguments passed to the SSH command for the host}
}
\value{
The \code{host} object.
}
\description{
Helps create a dockerfile for your own images running R
}
\details{
Dockerfiles are best practice when creating your own images.
}
\examples{
\dontrun{
docker_build(localhost, "/home/stuff/dockerfile" ,"new_image")
}
}
\seealso{
\href{https://docs.docker.com/engine/userguide/eng-image/dockerfile_best-practices/}{Best practices for writing Dockerfiles}

An example Dockerfile for \href{https://hub.docker.com/r/rocker/ropensci/~/dockerfile/}{rOpensci}

General R Docker images found at \href{https://github.com/rocker-org}{rocker-org}
}

