% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_gce.R
\name{docker_cmd.gce_instance}
\alias{docker_cmd.gce_instance}
\title{Docker S3 method for use with harbor package}
\usage{
\method{docker_cmd}{gce_instance}(host, cmd = NULL, args = NULL,
  docker_opts = NULL, capture_text = FALSE, ...)
}
\arguments{
\item{host}{The GCE instance}

\item{cmd}{The command to pass to docker}

\item{args}{arguments to the command}

\item{docker_opts}{options for docker}

\item{capture_text}{whether to return the output}

\item{...}{other arguments passed to \link{gce_ssh}}
}
\description{
Docker S3 method for use with harbor package
}
\details{
Instances launched in the \code{google-containers} image family automatically add your user to the docker group, 
  but for others you will need to run \code{sudo usermod -a -G docker ${USER}} and log out and back in.
}
