% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_docker.R
\name{docker_exec}
\alias{docker_exec}
\title{Execute a command within a docker container}
\usage{
docker_exec(host = localhost, container = NULL, ...)
}
\arguments{
\item{host}{A host object.}

\item{container}{The running container to execute within}

\item{...}{Other arguments passed to the SSH command for the host}
}
\value{
The \code{host} object.
}
\description{
Execute a command within a docker container
}
\examples{
\dontrun{
docker_exec(localhost, "container-id" ,"echo foo")
}
}

