% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_docker.R
\name{docker_inspect}
\alias{docker_inspect}
\title{Inspect one or more containers, given name(s) or ID(s).}
\usage{
docker_inspect(host = localhost, names = NULL, ...)
}
\arguments{
\item{host}{A host object.}

\item{names}{Names of the containers}

\item{...}{Other arguments passed to the SSH command for the host}
}
\value{
A list of lists, where each sublist represents one container. This is
  the output of `docker inspect` translated directly from raw JSON to an R
  object.
}
\description{
Inspect one or more containers, given name(s) or ID(s).
}
\examples{
\dontrun{
docker_run(localhost, "debian:testing", "echo foo", name = "harbor-test")
docker_inspect(localhost, "harbor-test")
}
}
\author{
Winston Change \email{winston@stdout.org}
}

