% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{gce_auth}
\alias{gce_auth}
\title{Defunct - Authenticate this session}
\usage{
gce_auth(new_user = FALSE, no_auto = FALSE)
}
\arguments{
\item{new_user}{If TRUE, reauthenticate via Google login screen}

\item{no_auto}{Will ignore auto-authentication settings if TRUE}
}
\value{
Invisibly, the token that has been saved to the session
}
\description{
No longer used.  Authenticate via downloading a JSON file and setting in your environment arguments instead.
}
