% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{gce_check_docker}
\alias{gce_check_docker}
\title{Check the docker logs of a container}
\usage{
gce_check_docker(instance, container)
}
\arguments{
\item{instance}{The instance running docker}

\item{container}{A running container to get logs of}
}
\value{
logs
}
\description{
Check the docker logs of a container
}

