% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpus.R
\name{gce_check_gpu}
\alias{gce_check_gpu}
\title{Check GPU installed ok}
\usage{
gce_check_gpu(vm)
}
\arguments{
\item{vm}{The instance to check}
}
\value{
The NVIDIA-SMI output via ssh
}
\description{
Check GPU installed ok
}
\seealso{
\url{https://cloud.google.com/compute/docs/gpus/add-gpus#verify-driver-install}

Other GPU instances: \code{\link{gce_list_gpus}},
  \code{\link{gce_vm_gpu}}
}
