% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gce_check_zone_op}
\alias{gce_check_zone_op}
\title{Wait for an operation to finish}
\usage{
gce_check_zone_op(job_name, wait = 3)
}
\arguments{
\item{wait}{Time in seconds between checks, default 3 seconds.}

\item{The}{job name}
}
\value{
The completed job object, invisibly
}
\description{
Will periodically check an operation until it is "DONE"
}

