% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{gce_container_addpackage}
\alias{gce_container_addpackage}
\title{Install a package to a running docker container on an instance}
\usage{
gce_container_addpackage(instance, container, cran_packages = NULL,
  github_packages = NULL, auth_token = NULL)
}
\arguments{
\item{instance}{The instance running the container}

\item{container}{The name of the running container}

\item{cran_packages}{A character vector of CRAN packages to be installed}

\item{github_packages}{A character vector of devtools packages to be installed}

\item{auth_token}{A github PAT token, if needed for private Github packages}
}
\value{
The instance
}
\description{
Add a package to an OpenCPU VM installed via \link{gce_vm_template}
}

