% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{gce_container_aptget}
\alias{gce_container_aptget}
\title{Install debain dependency libraries via apt-get in a container}
\usage{
gce_container_aptget(instance, container, call)
}
\arguments{
\item{instance}{The instance to install dependencies within}

\item{container}{The container to run \code{apt-get} within}

\item{call}{The dependency to install}
}
\value{
The instance
}
\description{
Install debain dependency libraries via apt-get in a container
}

