% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{gce_container_logs}
\alias{gce_container_logs}
\alias{gce_check_container}
\title{Check the docker logs of a container}
\usage{
gce_container_logs(instance, container)

gce_check_container(...)
}
\arguments{
\item{instance}{The instance running docker}

\item{container}{A running container to get logs of}

\item{...}{Arguments passed to \link{gce_container_logs}}
}
\value{
logs
}
\description{
Check the docker logs of a container
}
