% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{gce_containervm_create}
\alias{gce_containervm_create}
\title{Launch a container-VM image}
\usage{
gce_containervm_create(cloud_init, image_family = "gci-stable", ...)
}
\arguments{
\item{cloud_init}{template name or filepath to a cloud-init configuration file}

\item{image_family}{An image-family.  It must come from the \code{google-containers} family.}

\item{...}{Other arguments passed to \link{gce_vm_create}

Templates available are:

\itemize{
  \item rstudio An RStudio server docker image
  \item shiny A Shiny docker image
  \item opencpu An OpenCPU docker image
 }
 
 If not one of the above, then the function expects a filepath to a \href{cloud-init}{https://cloudinit.readthedocs.io/en/latest/topics/format.html} configuration file. 

\code{image_project} will be ignored if set, overriden to \code{google-containers}}
}
\value{
A zone operation
}
\description{
This lets you specify docker images when creating the VM
}

