% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disks.R
\name{gce_delete_disk}
\alias{gce_delete_disk}
\title{Deletes the specified persistent disk.}
\usage{
gce_delete_disk(disk, project = gce_get_global_project(),
  zone = gce_get_global_zone())
}
\arguments{
\item{disk}{Name of the persistent disk to delete}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}
}
\description{
Deleting a disk removes its data permanently and is irreversible.
}
\details{
However, deleting a disk does not delete any snapshots previously made from the disk. 
You must separately delete snapshots.


Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

