% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gce_delete_op}
\alias{gce_delete_op}
\title{Deletes the specified Operations resource.}
\usage{
gce_delete_op(operation)
}
\arguments{
\item{operation}{Name of the Operations resource to delete}
}
\value{
TRUE if successful
}
\description{
Deletes the specified Operations resource.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
