% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{gce_docker_build}
\alias{gce_docker_build}
\title{Build a dockerfile on a remote VM}
\usage{
gce_docker_build(dockerfile, build_name, ...)
}
\arguments{
\item{dockerfile}{The \code{Dockerfile} to build from}

\item{build_name}{The name of the Docker image to build}

\item{...}{Other parameters passed to \link{gce_vm_template}}
}
\description{
Instead of building locally or via SSH, launch a dedicated VM
}
\details{
This avoids using SSH by creating a dedicated VM to build your image, 
  and upload it to the Google Container Registry.

  
It uses the \code{builder} template available via \link{gce_vm_template}

If not otherwise specified as arguments, 
  the name of the VM will be \code{dockerbuild} and the predefined_type will be \code{n1-standard-1}

TODO: After uploading it will then delete the build VM
}
\examples{

\dontrun{

## A dockerfile that comes with the package to install googleAuthR packages
my_dockerfile <- system.file("dockerfiles",
                             "googleAuthR-verse", 
                             "Dockerfile", 
                             package = "googleComputeEngineR")
## build the dockerfile on remote VM                             
gce_docker_build(dockerfile = my_dockerfile, 
                 build_name = "test_docker_image")

}

}

