% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_admin.R
\name{gce_extract_projectzone}
\alias{gce_extract_projectzone}
\title{Extract zone and project from an instance object}
\usage{
gce_extract_projectzone(instance)
}
\arguments{
\item{instance}{The instance}
}
\value{
A list of $project and $zone
}
\description{
Extract zone and project from an instance object
}

