% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disks.R
\name{gce_get_disk}
\alias{gce_get_disk}
\title{Returns a specified persistent disk.}
\usage{
gce_get_disk(disk, project = gce_get_global_project(),
  zone = gce_get_global_zone())
}
\arguments{
\item{disk}{Name of the persistent disk to return}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}
}
\description{
Returns a specified persistent disk.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

