% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{gce_get_external_ip}
\alias{gce_get_external_ip}
\title{Get the external IP of an instance}
\usage{
gce_get_external_ip(instance, verbose = TRUE, ...)
}
\arguments{
\item{instance}{Name or instance object to find the external IP for}

\item{verbose}{Give a user message about the IP}

\item{...}{passed to \link{gce_get_instance}

This is a helper to extract the external IP of an instance}
}
\value{
The external IP
}
\description{
Get the external IP of an instance
}
