% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firewalls.R
\name{gce_get_firewall_rule}
\alias{gce_get_firewall_rule}
\title{Get a firewall rule}
\usage{
gce_get_firewall_rule(name, project = gce_get_global_project())
}
\arguments{
\item{name}{Name of the firewall rule}

\item{project}{The Google Cloud project}
}
\description{
Get a firewall rule of name specified
}

