% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{gce_get_image}
\alias{gce_get_image}
\title{Returns the specified image.}
\usage{
gce_get_image(project, image)
}
\arguments{
\item{project}{Project ID for this request}

\item{image}{Name of the image resource to return}
}
\description{
Returns the specified image.
}
\details{
Authentication scopes used by this function are: \itemize{ \item
https://www.googleapis.com/auth/cloud-platform \item
https://www.googleapis.com/auth/compute \item
https://www.googleapis.com/auth/compute.readonly }

You may want to use \link{gce_get_image_family} instead to ensure the most up to date image is used.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google
  Documentation}
}

