% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{gce_get_image_family}
\alias{gce_get_image_family}
\title{Returns the latest image that is part of an image family and is not deprecated.}
\usage{
gce_get_image_family(project, family)
}
\arguments{
\item{project}{Project ID for this request}

\item{family}{Name of the image family to search for}
}
\description{
Returns the latest image that is part of an image family and is not deprecated.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

