% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineType.R
\name{gce_get_machinetype}
\alias{gce_get_machinetype}
\title{Returns the specified machine type.}
\usage{
gce_get_machinetype(machineType, project = gce_get_global_project(),
  zone = gce_get_global_zone())
}
\arguments{
\item{machineType}{Name of the machine type to return}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}
}
\description{
Returns the specified machine type.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
