% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{gce_get_metadata}
\alias{gce_get_metadata}
\title{Extract metadata from an instance object}
\usage{
gce_get_metadata(instance, key = NULL)
}
\arguments{
\item{instance}{instance to get metadata from}

\item{key}{optional metadata key to filter metadata result}
}
\value{
data.frame $key and $value of metadata or NULL
}
\description{
Extract metadata from an instance object
}

