% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{gce_get_metadata_project}
\alias{gce_get_metadata_project}
\title{Get project wide metadata}
\usage{
gce_get_metadata_project(project = gce_global_project())
}
\arguments{
\item{project}{The project to get the project-wide metadata from}
}
\description{
Get project wide metadata
}
