% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{gce_get_network}
\alias{gce_get_network}
\title{Returns the specified network.}
\usage{
gce_get_network(network, project = gce_get_global_project())
}
\arguments{
\item{network}{Name of the network to return}

\item{project}{Project ID for this request}
}
\description{
Returns the specified network.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
