% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gce_get_op.gce_global_operation}
\alias{gce_get_op.gce_global_operation}
\title{Retrieves the specified global Operations resource.}
\usage{
\method{gce_get_op}{gce_global_operation}(operation,
  project = gce_get_global_project())
}
\arguments{
\item{operation}{Name of the Operations resource to return}

\item{project}{Project ID for this request}
}
\description{
Retrieves the specified global Operations resource.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

