% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gce_get_op}
\alias{gce_get_op}
\title{Retrieves the specified Operations resource.}
\usage{
gce_get_op(operation)
}
\arguments{
\item{operation}{Name of the Operations resource to return}
}
\description{
s3 method dispatcher
}
\details{

}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

