% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{gce_get_project}
\alias{gce_get_project}
\title{Returns the specified Project resource.}
\usage{
gce_get_project(project = gce_get_global_project())
}
\arguments{
\item{project}{Project ID for this request}
}
\description{
Returns the specified Project resource.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

