% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{gce_get_zone}
\alias{gce_get_zone}
\title{Returns the specified Zone resource. Get a list of available zones by making a list() request.}
\usage{
gce_get_zone(project, zone)
}
\arguments{
\item{project}{Project ID for this request}

\item{zone}{Name of the zone resource to return}
}
\description{
Returns the specified Zone resource. Get a list of available zones by making a list() request.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
