% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gce_get_zone_op}
\alias{gce_get_zone_op}
\title{Retrieves the specified zone-specific Operations resource.}
\usage{
gce_get_zone_op(operation, project = gce_get_global_project(),
  zone = gce_get_global_zone())
}
\arguments{
\item{operation}{Name of the Operations resource to return}

\item{project}{Project ID for this request}

\item{zone}{Name of the zone for this request}
}
\description{
Retrieves the specified zone-specific Operations resource.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

