% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{gce_global_zone}
\alias{gce_global_zone}
\title{Set global zone name}
\usage{
gce_global_zone(zone)
}
\arguments{
\item{zone}{zone name you want this session to use by default, or a zone object}
}
\value{
The zone name (invisibly)
}
\description{
Set a zone name used for this R session
}
\details{
This sets a zone to a global environment value so you don't need to
supply the zone argument to other API calls.
}
