% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disks.R
\name{gce_list_disks_all}
\alias{gce_list_disks_all}
\title{Retrieves an aggregated list of persistent disks across all zones.}
\usage{
gce_list_disks_all(filter = NULL, maxResults = NULL,
  pageToken = NULL, project = gce_get_global_project())
}
\arguments{
\item{filter}{Sets a filter expression for filtering listed resources, in the form filter={expression}}

\item{maxResults}{The maximum number of results per page that should be returned}

\item{pageToken}{Specifies a page token to use}

\item{project}{Project ID for this request}
}
\description{
Retrieves an aggregated list of persistent disks across all zones.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
