% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gce_list_zone_op}
\alias{gce_list_zone_op}
\title{Retrieves a list of Operation resources contained within the specified zone.}
\usage{
gce_list_zone_op(filter = NULL, maxResults = NULL, pageToken = NULL,
  project = gce_get_global_project(), zone = gce_get_global_zone())
}
\arguments{
\item{filter}{Sets a filter expression for filtering listed resources, in the form filter={expression}}

\item{maxResults}{The maximum number of results per page that should be returned}

\item{pageToken}{Specifies a page token to use}

\item{project}{Project ID for this request}

\item{zone}{Name of the zone for request}
}
\description{
Retrieves a list of Operation resources contained within the specified zone.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

