% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{gce_list_zones}
\alias{gce_list_zones}
\title{Retrieves the list of Zone resources available to the specified project.}
\usage{
gce_list_zones(project, filter = NULL, maxResults = NULL,
  pageToken = NULL)
}
\arguments{
\item{project}{Project ID for this request}

\item{filter}{Sets a filter expression for filtering listed resources, in the form filter={expression}}

\item{maxResults}{The maximum number of results per page that should be returned}

\item{pageToken}{Specifies a page token to use}
}
\description{
Retrieves the list of Zone resources available to the specified project.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
