% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disks.R
\name{gce_make_boot_disk}
\alias{gce_make_boot_disk}
\title{Make a boot disk for attachment to an instance}
\usage{
gce_make_boot_disk(diskName = NULL, diskSizeGb = NULL,
  diskType = NULL, sourceImage = NULL,
  sourceImageEncryptionKey = NULL)
}
\arguments{
\item{diskName}{Specifies the disk name}

\item{diskSizeGb}{Specifies the size of the disk in base-2 GB}

\item{diskType}{Specifies the disk type to use to create the instance}

\item{sourceImage}{The source image used to create this disk}

\item{sourceImageEncryptionKey}{The customer-supplied encryption key of the source image}
}
\value{
AttachedDiskInitializeParams object
}
\description{
Make a boot disk for attachment to an instance
}
\details{
Specifies the parameters for a new disk that will be created alongside the new instance. 

Use initialization parameters to create boot disks or local SSDs attached to the new instance.

This property is mutually exclusive with the source property; you can only define one or the other, but not both.
}
\concept{AttachedDiskInitializeParams functions}
