% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firewalls.R
\name{gce_make_firewall_rule}
\alias{gce_make_firewall_rule}
\title{Add one firewall rule to the network}
\usage{
gce_make_firewall_rule(name, protocol, ports, sourceRanges = NULL,
  sourceTags = NULL, project = gce_get_global_project())
}
\arguments{
\item{name}{Name of the firewall rule}

\item{protocol}{Protocol such as \code{tcp, udp, icmp, esp, ah, sctp} or IP protocol number.}

\item{ports}{Port numbers to open}

\item{sourceRanges}{From where to accept connections.  If \code{NULL} then will default to \code{0.0.0.0/0} (everywhere)}

\item{sourceTags}{A list of instance tags this rule applies to. One or both of \code{sourceRanges} and \code{sourceTags} may be set.}

\item{project}{The Google Cloud project}
}
\value{
A global operation object
}
\description{
Use this to create firewall rules to apply to the network settings.  
  Most commonly this is to setup web access (port 80 and 443)
}
\section{sourceRanges and/or sourceTags}{


If both properties are set, 
  an inbound connection is allowed if the range or the tag of the source matches the 
  sourceRanges OR matches the sourceTags property; the connection does not need to match both properties.
}

\examples{

\dontrun{

  gce_make_firewall_rule("allow-http", protocol = "tcp", ports = 80)
  gce_make_firewall_rule("allow-https", protocol = "tcp", ports = 443)
  
}

}
\seealso{
API Documentation \url{https://cloud.google.com/compute/docs/reference/latest/firewalls/insert}

Other firewall functions: \code{\link{gce_delete_firewall_rule}},
  \code{\link{gce_get_firewall_rule}},
  \code{\link{gce_list_firewall_rules}},
  \code{\link{gce_make_firewall_webports}}
}
