% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firewalls.R
\name{gce_make_firewall_webports}
\alias{gce_make_firewall_webports}
\title{Make HTTP and HTTPS firewall rules}
\usage{
gce_make_firewall_webports(project = gce_get_global_project())
}
\arguments{
\item{project}{The project the firewall will open for}
}
\value{
Vector of the firewall objects
}
\description{
Do the common use case of opening HTTP and HTTPS ports
}
\details{
This will invoke \link{gce_make_firewall_rule} and look for the rules named \code{allow-http} and \code{allow-https}.
If not present, it will create them.
}
\seealso{
Other firewall functions: \code{\link{gce_delete_firewall_rule}},
  \code{\link{gce_get_firewall_rule}},
  \code{\link{gce_list_firewall_rules}},
  \code{\link{gce_make_firewall_rule}}
}
