% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{gce_make_image_source_url}
\alias{gce_make_image_source_url}
\title{Make initial disk image object}
\usage{
gce_make_image_source_url(image_project, image = NULL, family = NULL)
}
\arguments{
\item{image_project}{Project ID of where the image lies}

\item{image}{Name of the image resource to return}

\item{family}{Name of the image family to search for}
}
\value{
The selfLink of the image object
}
\description{
Make initial disk image object
}
