% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineType.R
\name{gce_make_machinetype_url}
\alias{gce_make_machinetype_url}
\title{Construct a machineType URL}
\usage{
gce_make_machinetype_url(predefined_type = NULL, cpus = NULL,
  memory = NULL, zone = gce_get_global_zone())
}
\arguments{
\item{predefined_type}{A predefined machine type from \link{gce_list_machinetype}}

\item{cpus}{If not defining \code{predefined_type}, the number of CPUs}

\item{memory}{If not defining \code{predefined_type}, amount of memory}

\item{zone}{zone for URL}
}
\value{
A url for use in instance creation
}
\description{
Construct a machineType URL
}
\details{
\code{cpus} must be in multiples of 2 up to 32
\code{memory} must be in multiples of 256
}
