% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{gce_make_network}
\alias{gce_make_network}
\title{Make a network interface for instance creation}
\usage{
gce_make_network(network = "default", project = gce_get_global_project())
}
\arguments{
\item{network}{Name of network resource}

\item{project}{Project ID for this request}
}
\value{
A Network object
}
\description{
Make a network interface for instance creation
}

