% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{gce_metadata_env}
\alias{gce_metadata_env}
\title{Turn metadata into an environment argument}
\usage{
gce_metadata_env(key)
}
\arguments{
\item{key}{The metadata key.  Pass "" to list the keys}
}
\value{
The metadata key value, if successful
}
\description{
This turns instance metadata into an environment argument R (and other software) can see.  
  Only works on a running instance.
}
