% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencpu.R
\name{gce_opencpu_addpackage}
\alias{gce_opencpu_addpackage}
\title{Add a package to a running OpenCPU instance}
\usage{
gce_opencpu_addpackage(instance, cran_packages = NULL,
  github_packages = NULL, auth_token = NULL)
}
\arguments{
\item{instance}{The instance running Shiny}

\item{cran_packages}{A character vector of CRAN packages to be installed}

\item{github_packages}{A character vector of devtools packages to be installed}

\item{auth_token}{A github PAT token, if needed for private Github packages}
}
\value{
The instance
}
\description{
Add a package to an OpenCPU VM installed via \link{gce_vm_template}
}

