% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{gce_rstudio_password}
\alias{gce_rstudio_password}
\title{Changes password for a user on RStudio container}
\usage{
gce_rstudio_password(instance, username, password, container = "rstudio")
}
\arguments{
\item{instance}{An instance with RStudio installed via \link{gce_vm_template}}

\item{username}{The user to chnage the password for}

\item{password}{The user password}

\item{container}{The rstudio container to add the user to}
}
\value{
The instance
}
\description{
RStudio has users based on unix user accounts
}

