% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{gce_save_container}
\alias{gce_save_container}
\title{Save the docker container to the private Google Container Registry}
\usage{
gce_save_container(instance, container_name, template_name = "rstudio",
  container_url = "gcr.io", project = gce_get_global_project())
}
\arguments{
\item{instance}{The VM to run within}

\item{container_name}{The name for the saved container}

\item{template_name}{The template you used to start the instance.}

\item{container_url}{The URL of where to save container}

\item{project}{Project ID for this request, default as set by \link{gce_get_global_project}

This will only work on the Google Container optimised containers of image_family google_containers.
Otherwise you will need to get a container authentication yourself (for now)

It will start the push but it may take a long time to finish, espeically the first time, 
  this function will return whilst waiting but don't turn off the VM until its finished.}
}
\value{
TRUE if commands finish
}
\description{
Saves the docker image to your projects Google Cloud Storage.
}

