% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{gce_schedule_build_script}
\alias{gce_schedule_build_script}
\title{Schedule a script upon a VM}
\usage{
gce_schedule_build_script(script, vm, schedule, ...)
}
\arguments{
\item{script}{The script to schedule}

\item{vm}{A VM object to schedule the script upon}

\item{schedule}{The schedule you want to run via cron}

\item{...}{Other arguments pass to \code{containeRit::dockerfile}}
}
\description{
Utility function to start a VM, upload your script and set it up on a schedule
}
