% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_admin.R
\name{gce_set_machinetype}
\alias{gce_set_machinetype}
\title{Changes the machine type for a stopped instance to the machine type specified in the request.}
\usage{
gce_set_machinetype(machineType, project, zone, instance)
}
\arguments{
\item{machineType}{Full or partial URL of the machine type resource}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}

\item{instance}{Name of the instance scoping this request}
}
\description{
Changes the machine type for a stopped instance to the machine type specified in the request.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstancesSetMachineTypeRequest.functions: \code{\link{InstancesSetMachineTypeRequest}}
}

