% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min-cpu-platform.R
\name{gce_set_mincpuplatform}
\alias{gce_set_mincpuplatform}
\title{Set a minCPU platform on a stopped instance}
\usage{
gce_set_mincpuplatform(instance, minCpuPlatform)
}
\arguments{
\item{instance}{The (stopped) instance to set a minimum CPU platform upon}

\item{minCpuPlatform}{The platform to set}
}
\description{
Set a minCPU platform on a stopped instance
}
