% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{gce_shiny_addpackage}
\alias{gce_shiny_addpackage}
\title{Add a package to a running Shiny instance}
\usage{
gce_shiny_addpackage(instance, cran_packages = NULL, github_packages = NULL,
  auth_token = NULL)
}
\arguments{
\item{instance}{The instance running Shiny}

\item{cran_packages}{A character vector of CRAN packages to be installed}

\item{github_packages}{A character vector of devtools packages to be installed}

\item{auth_token}{A github PAT token, if needed for private Github packages}
}
\value{
The instance
}
\description{
Add a package to a Shiny VM installed via \link{gce_vm_template}
}

