% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{gce_shiny_listapps}
\alias{gce_shiny_listapps}
\title{List shiny apps on the instance}
\usage{
gce_shiny_listapps(instance)
}
\arguments{
\item{instance}{Instance with Shiny apps installed}
}
\value{
character vector
}
\description{
List shiny apps on the instance
}

