% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{gce_shiny_logs}
\alias{gce_shiny_logs}
\title{Get the latest shiny logs for a shinyapp}
\usage{
gce_shiny_logs(instance, shinyapp = NULL)
}
\arguments{
\item{instance}{Instance with Shiny app installed}

\item{shinyapp}{Name of shinyapp to see logs for. If NULL will return general shiny logs}
}
\value{
log printout
}
\description{
Get the latest shiny logs for a shinyapp
}
