% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{gce_ssh_browser}
\alias{gce_ssh_browser}
\title{Open a cloud SSH browser for an instance}
\usage{
gce_ssh_browser(instance, project = gce_get_global_project(),
  zone = gce_get_global_zone())
}
\arguments{
\item{instance}{Name of the instance resource}

\item{project}{Project ID for this request, default as set by \link{gce_get_global_project}}

\item{zone}{The name of the zone for this request, default as set by \link{gce_get_global_zone}}
}
\value{
Opens a browser window to the SSH session, returns the SSH URL.
}
\description{
This will open an SSH from the browser session if \code{getOption("browser")} is not NULL
}
\details{
You will need to login the first time with an email that has access to the instance.
}
\seealso{
\url{https://cloud.google.com/compute/docs/ssh-in-browser}
}

