% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{gce_ssh_setup}
\alias{gce_ssh_setup}
\title{Setup a SSH connection with GCE from a new SSH key-pair}
\usage{
gce_ssh_setup(key.pub, instance, project = gce_get_global_project(),
  zone = gce_get_global_zone())
}
\arguments{
\item{key.pub}{The contents of the \code{~/.ssh/my-ssh-key.pub} generated by the cat command}

\item{instance}{Name of the instance of run ssh command upon}

\item{project}{Project ID for this request, default as set by \link{gce_get_global_project}}

\item{zone}{The name of the zone for this request, default as set by \link{gce_get_global_zone}}
}
\description{
You will need to generate a new SSH key-pair if you have not connected to the instance before. Instructions for this can be found here: \url{https://cloud.google.com/compute/docs/instances/connecting-to-instance}.  Once you have generated run this function once to initiate setup
}
\seealso{
\url{https://cloud.google.com/compute/docs/instances/connecting-to-instance}

Other ssh functions: \code{\link{gce_ssh_browser}},
  \code{\link{gce_ssh}}
}

