% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{gce_vm_create}
\alias{gce_vm_create}
\title{Creates an instance resource in the specified project using the data included in the request.}
\usage{
gce_vm_create(name, predefined_type, cpus, memory, canIpForward = NULL,
  description = NULL, disks = NULL, metadata = NULL,
  networkInterfaces = NULL, scheduling = NULL, serviceAccounts = NULL,
  tags = NULL, project = gce_get_global_project(),
  zone = gce_get_global_zone())
}
\arguments{
\item{name}{The name of the resource, provided by the client when initially creating the resource}

\item{predefined_type}{A predefined machine type from \link{gce_list_machinetype}}

\item{cpus}{If not defining \code{predefined_type}, the number of CPUs}

\item{memory}{If not defining \code{predefined_type}, amount of memory}

\item{canIpForward}{Allows this instance to send and receive packets with non-matching destination or source IPs}

\item{description}{An optional description of this resource}

\item{disks}{Array of disks associated with this instance}

\item{metadata}{The metadata key/value pairs assigned to this instance}

\item{networkInterfaces}{An array of configurations for this interface}

\item{scheduling}{Scheduling options for this instance}

\item{serviceAccounts}{A list of service accounts, with their specified scopes, authorized for this instance}

\item{tags}{A list of tags to apply to this instance}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}
}
\description{
Creates an instance resource in the specified project using the data included in the request.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

\code{cpus} must be in multiples of 2 up to 32
\code{memory} must be in multiples of 256
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other Instance functions: \code{\link{Instance}}
}

