% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{gce_vm_delete}
\alias{gce_vm_delete}
\title{Deletes the specified Instance resource.}
\usage{
gce_vm_delete(project, zone, instance)
}
\arguments{
\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}

\item{instance}{Name of the instance resource to delete}
}
\description{
Deletes the specified Instance resource.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

