% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{gce_vm_scheduler}
\alias{gce_vm_scheduler}
\title{Create or start a scheduler VM}
\usage{
gce_vm_scheduler(vm_name = "scheduler", ...)
}
\arguments{
\item{vm_name}{The name of the VM scheduler to create or return}

\item{...}{Arguments passed on to \code{gce_vm}
\describe{
  \item{name}{The name of the instance}
  \item{open_webports}{If TRUE, will open firewall ports 80 and 443 if not open already}
  \item{project}{Project ID for this request}
  \item{zone}{The name of the zone for this request}
}}
}
\value{
A VM object
}
\description{
This starts up a VM with cron and docker installed that can be used to schedule scripts
}
\seealso{
Other scheduler functions: \code{\link{gce_schedule_docker}}
}
