% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gce_wait}
\alias{gce_wait}
\title{Wait for an operation to finish}
\usage{
gce_wait(operation, wait = 3, verbose = TRUE, timeout_tries = 50)
}
\arguments{
\item{operation}{The operation object}

\item{wait}{Time in seconds between checks, default 3 seconds.}

\item{verbose}{Whether to give user feedback}

\item{timeout_tries}{Number of times to wait}
}
\value{
The completed job object, invisibly
}
\description{
Will periodically check an operation until its status is \code{DONE}
}
