% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{gcs_get_global_project}
\alias{gcs_get_global_project}
\title{Get global project name}
\usage{
gcs_get_global_project()
}
\value{
Project name
}
\description{
Project name set this session to use by default
}
\details{
Set the project name via \link{gcs_global_project}
}

