% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{gcs_get_global_zone}
\alias{gcs_get_global_zone}
\title{Get global zone name}
\usage{
gcs_get_global_zone()
}
\value{
zone name
}
\description{
zone name set this session to use by default
}
\details{
Set the zone name via \link{gcs_global_zone}
}

