% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{get_cloud_init_file}
\alias{get_cloud_init_file}
\title{create the cloud_init file to upload}
\usage{
get_cloud_init_file(template, username = NULL, password = NULL,
  dynamic_image = NULL)
}
\description{
create the cloud_init file to upload
}
\keyword{internal}
