% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{get_dockerfile}
\alias{get_dockerfile}
\title{Get Dockerfile templates}
\usage{
get_dockerfile(dockerfile_folder)
}
\arguments{
\item{dockerfile_folder}{}
}
\value{
file location
}
\description{
This gets the file location of available Dockerfile examples
}

