% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{get_dockerfolder}
\alias{get_dockerfolder}
\title{Get Dockerfolder of templates}
\usage{
get_dockerfolder(dockerfile_folder)
}
\arguments{
\item{dockerfile_folder}{The folder containing \code{Dockerfile}}
}
\value{
file location
}
\description{
This gets the folder location of available Dockerfile examples
}
