% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleComputeEngineR.R
\docType{package}
\name{googleComputeEngineR}
\alias{googleComputeEngineR}
\alias{googleComputeEngineR-package}
\title{Working with Google Compute Engine from R}
\description{
See demos and examples at the \href{googleComputeEngineR website}{https://cloudyr.github.io/googleComputeEngineR/}.
}
