% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{Instance}
\alias{Instance}
\title{Instance Object}
\usage{
Instance(canIpForward = NULL, description = NULL, disks = NULL,
  machineType = NULL, metadata = NULL, name = NULL,
  networkInterfaces = NULL, scheduling = NULL, serviceAccounts = NULL,
  tags = NULL)
}
\arguments{
\item{canIpForward}{Allows this instance to send and receive packets with non-matching destination or source IPs}

\item{description}{An optional description of this resource}

\item{disks}{Array of disks associated with this instance}

\item{machineType}{Full or partial URL of the machine type resource to use for this instance, in the format: zones/zone/machineTypes/machine-type}

\item{metadata}{The metadata key/value pairs assigned to this instance}

\item{name}{The name of the resource, provided by the client when initially creating the resource}

\item{networkInterfaces}{An array of configurations for this interface}

\item{scheduling}{Scheduling options for this instance}

\item{serviceAccounts}{A list of service accounts, with their specified scopes, authorized for this instance}

\item{tags}{A list of tags to apply to this instance}
}
\value{
Instance object
}
\description{
Instance Object
}
\details{
An Instance resource.
}
\seealso{
Other Instance.functions: \code{\link{gce_vm_create}}
}
\keyword{internal}

