% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_admin.R
\name{InstancesSetMachineTypeRequest}
\alias{InstancesSetMachineTypeRequest}
\title{InstancesSetMachineTypeRequest Object}
\usage{
InstancesSetMachineTypeRequest(machineType = NULL)
}
\arguments{
\item{machineType}{Full or partial URL of the machine type resource}
}
\value{
InstancesSetMachineTypeRequest object
}
\description{
InstancesSetMachineTypeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstancesSetMachineTypeRequest functions: \code{\link{gce_set_machinetype}}
}
\keyword{internal}

