% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{is.gce_global_operation}
\alias{is.gce_global_operation}
\title{Check if is a gce_global_operation}
\usage{
is.gce_global_operation(x)
}
\arguments{
\item{x}{The object to test if class \code{gce_global_operation}}
}
\value{
TRUE or FALSE
}
\description{
Check if is a gce_global_operation
}
