% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_admin.R
\name{is.gce_instance}
\alias{is.gce_instance}
\title{Check if is gce_instance}
\usage{
is.gce_instance(x)
}
\arguments{
\item{x}{The object to test if class \code{gce_instance}}
}
\value{
TRUE or FALSE
}
\description{
Check if is gce_instance
}

