% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{Metadata}
\alias{Metadata}
\title{Metadata Object}
\usage{
Metadata(items)
}
\arguments{
\item{items}{A named list of key = value pairs}
}
\value{
Metadata object
}
\description{
Metadata Object
}
\keyword{internal}

