% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud-init.R
\name{read_cloud_init_file}
\alias{read_cloud_init_file}
\title{create the cloud_init file to upload}
\usage{
read_cloud_init_file(template)
}
\description{
create the cloud_init file to upload
}
\keyword{internal}
