% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup-scripts.R
\name{read_shell_startup_file}
\alias{read_shell_startup_file}
\title{create the shell file to upload}
\usage{
read_shell_startup_file(template)
}
\description{
create the shell file to upload
}
\keyword{internal}
