% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.R
\name{as.cluster.gce_instance}
\alias{as.cluster.gce_instance}
\title{Create a future cluster for GCE objects}
\usage{
\method{as.cluster}{gce_instance}(x, user = gce_get_global_ssh_user(),
  project = gce_get_global_project(), zone = gce_get_global_zone(),
  rshopts = ssh_options(), ..., recursive = FALSE)
}
\arguments{
\item{x}{The instance to make a future cluster}

\item{user}{Username used in SSH}

\item{project}{The GCE project}

\item{zone}{The GCE zone}

\item{rshopts}{Options for the SSH}

\item{...}{Other arguments passed to makeDockerClusterPSOCK}

\item{recursive}{Not used.}
}
\value{
A \code{cluster} object.
}
\description{
S3 method for \code{\link[future:as.cluster]{as.cluster}()} in the \pkg{future} package.
}
\examples{
\donttest{\dontrun{
vm <- gce_vm_template("r-base", name = "future", predefined_type = "f1-micro")
plan(cluster, workers = as.cluster(vm))
x \%<-\% { Sys.getinfo() }
print(x)
}}

}

