% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_container.R
\name{as.container}
\alias{as.container}
\title{Coerce an object into a container object.}
\usage{
as.container(x, host = localhost)
}
\arguments{
\item{x}{An object to coerce}

\item{host}{A docker host

A container object represents a Docker container on a host.}
}
\description{
Coerce an object into a container object.
}
\author{
Winston Change \email{winston@stdout.org}
}

