% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_admin.R
\name{as.gce_instance_name}
\alias{as.gce_instance_name}
\title{Get the instance name(s) if passed instance(s)}
\usage{
as.gce_instance_name(x)
}
\arguments{
\item{A}{list or a single instance}
}
\description{
Get the instance name(s) if passed instance(s)
}
\keyword{internal}

