% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{auth_email}
\alias{auth_email}
\title{Get auth email
If it includes '@' then assume the email, otherwise an environment file}
\usage{
auth_email(source)
}
\arguments{
\item{source}{where the email comes from}
}
\description{
Get auth email
If it includes '@' then assume the email, otherwise an environment file
}
\keyword{internal}

