% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_container.R
\name{container_logs}
\alias{container_logs}
\title{Retrieve logs for a container.}
\usage{
container_logs(container, timestamps = FALSE, follow = FALSE)
}
\arguments{
\item{container}{A container object}

\item{timestamps}{Show timestamps.}

\item{follow}{Follow log output as it is happening.}
}
\description{
Retrieve logs for a container.
}
\examples{
\dontrun{
container_rm(con)
}
}
\author{
Winston Change \email{winston@stdout.org}
}

