% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_container.R
\name{container_running}
\alias{container_running}
\title{Report whether a container is currently running.}
\usage{
container_running(container)
}
\arguments{
\item{container}{A container object}
}
\description{
Report whether a container is currently running.
}
\examples{
\dontrun{
container_running(con)
}
}
\author{
Winston Change \email{winston@stdout.org}
}

