% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_containers.R
\name{containers}
\alias{containers}
\title{Get list of all containers on a host.}
\usage{
containers(host = localhost, ...)
}
\arguments{
\item{host}{A host object.}

\item{...}{Other arguments passed to the SSH command for the host}
}
\description{
Get list of all containers on a host.
}
\author{
Winston Change \email{winston@stdout.org}
}

