% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_docker.R
\name{docker_pull}
\alias{docker_pull}
\title{Pull a docker image onto a host.}
\usage{
docker_pull(host = localhost, image, ...)
}
\arguments{
\item{host}{A host object.}

\item{image}{The docker image to pull e.g. \code{rocker/rstudio}}

\item{...}{Other arguments passed to the SSH command for the host}
}
\value{
The \code{host} object.
}
\description{
Pull a docker image onto a host.
}
\examples{
\dontrun{
docker_pull(localhost, "debian:testing")
}
}
\author{
Winston Change \email{winston@stdout.org}
}

