% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_admin.R
\name{gce_check_ssh}
\alias{gce_check_ssh}
\title{Calls API for the current SSH settings for an instance}
\usage{
gce_check_ssh(instance)
}
\arguments{
\item{instance}{An instance to check}
}
\value{
A data.frame of SSH users and public keys
}
\description{
Calls API for the current SSH settings for an instance
}
