% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firewalls.R
\name{gce_get_firewall_rule}
\alias{gce_get_firewall_rule}
\title{Get a firewall rule}
\usage{
gce_get_firewall_rule(name, project = gce_get_global_project())
}
\arguments{
\item{name}{Name of the firewall rule}

\item{project}{The Google Cloud project}
}
\description{
Get a firewall rule of name specified
}
\seealso{
API Documentation \url{https://cloud.google.com/compute/docs/reference/latest/firewalls/get}

Other firewall functions: \code{\link{gce_delete_firewall_rule}},
  \code{\link{gce_list_firewall_rules}},
  \code{\link{gce_make_firewall_rule}},
  \code{\link{gce_make_firewall_webports}}
}
\concept{firewall functions}
