% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_admin.R
\name{gce_get_instance}
\alias{gce_get_instance}
\title{Returns the specified Instance resource.}
\usage{
gce_get_instance(instance, project = gce_get_global_project(),
  zone = gce_get_global_zone())
}
\arguments{
\item{instance}{Name of the instance resource}

\item{project}{Project ID for this request, default as set by \link{gce_get_global_project}}

\item{zone}{The name of the zone for this request, default as set by \link{gce_get_global_zone}}
}
\description{
Returns the specified Instance resource.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
