% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gce_get_op.gce_zone_operation}
\alias{gce_get_op.gce_zone_operation}
\title{Retrieves the specified zone-specific Operations resource.}
\usage{
\method{gce_get_op}{gce_zone_operation}(operation)
}
\arguments{
\item{operation}{Name of the Operations resource to return}
}
\description{
Retrieves the specified zone-specific Operations resource.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
