% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gce_get_op}
\alias{gce_get_op}
\title{Retrieves the specified Operations resource.}
\usage{
gce_get_op(operation = .Last.value)
}
\arguments{
\item{operation}{Name of the Operations resource to return}
}
\description{
s3 method dispatcher
}
\details{
S3 Methods for classes
\itemize{
  \item gce_get_op.gce_zone_operation
  \item gce_get_op.gce_global_operation
  \item gce_get_op.gce_region_operation
 }
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
