% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{gce_install_packages_container}
\alias{gce_install_packages_container}
\title{Install packages in a instance's running container}
\usage{
gce_install_packages_container(instance, container, cran_packages = NULL,
  github_packages = NULL, auth_token = devtools::github_pat())
}
\arguments{
\item{instance}{The instance running the container}

\item{container}{The container name running R to install packages within}

\item{cran_packages}{A character vector of CRAN packages to be installed}

\item{github_packages}{A character vector of devtools packages to be installed}

\item{auth_token}{A Github PAT for private repos if needed}
}
\value{
TRUE if successful
}
\description{
Install packages in a instance's running container
}

