% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{gce_shiny_addapp}
\alias{gce_shiny_addapp}
\title{Add Shiny app to a Shiny template instance}
\usage{
gce_shiny_addapp(instance, shinyapp = ".", ...)
}
\arguments{
\item{instance}{The instance running Shiny}

\item{shinyapp}{The folder container the local Shiny app}

\item{...}{Other arguments passed to \link{gce_ssh_upload}}
}
\value{
The instance
}
\description{
Add a local shiny app to a running Shiny VM installed via \link{gce_vm_template}
}

