% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{gce_ssh_browser}
\alias{gce_ssh_browser}
\title{Open a cloud SSH browser for an instance}
\usage{
gce_ssh_browser(instance)
}
\arguments{
\item{instance}{the instance resource}
}
\value{
Opens a browser window to the SSH session, returns the SSH URL.
}
\description{
This will open an SSH from the browser session if \code{getOption("browser")} is not NULL
}
\details{
You will need to login the first time with an email that has access to the instance.
}
\seealso{
\url{https://cloud.google.com/compute/docs/ssh-in-browser}

Other ssh functions: \code{\link{gce_ssh_setup}},
  \code{\link{gce_ssh}}
}

