% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{gce_startup_logs}
\alias{gce_startup_logs}
\title{Get startup script logs}
\usage{
gce_startup_logs(instance, type = c("shell", "cloud-config", "nginx"))
}
\arguments{
\item{instance}{The instance to get startup script logs from}

\item{type}{The type of log to run

Will use SSH so that needs to be setup}
}
\description{
Get startup script logs
}
