% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{gce_vm_container}
\alias{gce_vm_container}
\title{Launch a container-VM image}
\usage{
gce_vm_container(file = NULL, cloud_init = NULL, shell_script = NULL,
  image_family = "cos-stable", image_project = "cos-cloud", ...)
}
\arguments{
\item{file}{file location of a valid cloud-init or shell_script file. 
One of \code{file} or \code{cloud_init} or \code{shell_script} must be supplied}

\item{cloud_init}{contents of a cloud-init file, for example read via \code{readChar(file, nchars = 32768)}}

\item{shell_script}{contents of a shell_script file, for example read via \code{readChar(file, nchars = 32768)}}

\item{image_family}{An image-family.  It must come from the \code{image_project} family.}

\item{image_project}{An image-project, where the image-family resides.}

\item{...}{Other arguments passed to \link{gce_vm_create}}
}
\value{
A zone operation
}
\description{
This lets you specify docker images when creating the VM.  These are a special class of Google instances that are setup for running Docker containers.
}
\details{
\code{file} expects a filepath to a \href{cloud-init}{https://cloudinit.readthedocs.io/en/latest/topics/format.html} configuration file or a valid bash script e.g. has \code{!#/bin/} or \code{#cloud-config} at top of file.

\code{image_project} will be ignored if set, overriden to \code{cos-cloud}.  
If you want to set it then use the \link{gce_vm_create} function directly that this function wraps with some defaults.
}
\seealso{
\url{https://cloud.google.com/container-optimized-os/docs/how-to/create-configure-instance} - \href{https://www.freedesktop.org/software/systemd/man/systemd.service.html}{help using cloud-init files}
}
