% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion-prevention.R
\name{gce_vm_deletion_protection}
\alias{gce_vm_deletion_protection}
\title{Toggle deletion protection for existing instances}
\usage{
gce_vm_deletion_protection(instance, cmd = c("status", "true", "false"),
  projectId = gce_get_global_project(), zone = gce_get_global_zone())
}
\arguments{
\item{instance}{The vm to work with its deletion protection}

\item{cmd}{Whether to get the status, or toggle "true" or "false" 
on deletion protection for this VM}

\item{projectId}{The projectId}

\item{zone}{The zone}
}
\description{
Toggle deletion protection for existing instances
}
