% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms.R
\name{gce_vm_reset}
\alias{gce_vm_reset}
\title{Performs a hard reset on the instance.}
\usage{
gce_vm_reset(instance, project = gce_get_global_project(),
  zone = gce_get_global_zone())
}
\arguments{
\item{instance}{Name of the instance resource, or an instance object e.g. from \link{gce_get_instance}}

\item{project}{Project ID for this request, default as set by \link{gce_get_global_project}}

\item{zone}{The name of the zone for this request, default as set by \link{gce_get_global_zone}}
}
\description{
Performs a hard reset on the instance.
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
