% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_admin.R
\name{Instance}
\alias{Instance}
\title{Instance Object}
\usage{
Instance(name = NULL, machineType = NULL, canIpForward = NULL,
  description = NULL, disks = NULL, metadata = NULL,
  networkInterfaces = NULL, scheduling = NULL, serviceAccounts = NULL,
  tags = NULL)
}
\arguments{
\item{name}{The name of the resource, provided by the client when initially creating the resource}

\item{machineType}{Full or partial URL of the machine type resource to use for this instance, in the format: \code{zones/zone/machineTypes/machine-type}}

\item{canIpForward}{Allows this instance to send and receive packets with non-matching destination or source IPs}

\item{description}{An optional description of this resource}

\item{disks}{The source image used to create this disk}

\item{metadata}{A named list of metadata key/value pairs assigned to this instance}

\item{networkInterfaces}{An array of configurations for this interface}

\item{scheduling}{Scheduling options for this instance}

\item{serviceAccounts}{A list of service accounts, with their specified scopes, authorized for this instance}

\item{tags}{A list of tags to apply to this instance}
}
\value{
Instance object
}
\description{
Instance Object
}
\details{
An Instance resource.
}
\keyword{internal}

