% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harbor_container.R
\docType{data}
\name{localhost}
\alias{localhost}
\title{An object representing the current computer that R is running on.}
\format{An object of class \code{localhost} (inherits from \code{host}) of length 0.}
\usage{
localhost
}
\description{
An object representing the current computer that R is running on.
}
\keyword{datasets}

